Rename include directory for ROS-related headers
