+++ launch/nodelet.launch
@@ -13,7 +13,6 @@
   <param name="tag_size_cm" value="$(arg tag_size_cm)" /> 
   <node pkg="nodelet" name="april_tag" type="nodelet"
         args="load april_tag/AprilTagNodelet $(arg manager)" output="screen">
-    <remap from="/camera/image_raw" to="$(arg camera_ns)/image_raw" />
     <param name="camera_ns" value="$(arg camera_ns)" />
 
     <param name="focal_length_x_px" value="$(arg focal_length_x_px)" />
